tts-services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
soundfile==0.12.1
requests==2.31.0
aiofiles==23.2.1
aiohttp==3.9.1
edge-tts==6.1.9
openai==1.3.0
python-dotenv==1.0.0
